Database addition and modification
ID datatype changed from integer to bigInteger and added some database regarding accounting and money

+++ klugee-main-thing/database/migrations/2021_11_27_085211_create_cutis_table.php
@@ -15,7 +15,7 @@ class CreateCutisTable extends Migration
     {
         Schema::create('cutis', function (Blueprint $table) {
             $table->id();
-            $table->integer("id_teacher");
+            $table->bigInteger("id_teacher");
             $table->date("date_begin");
             $table->date("date_end");
             $table->boolean("approved");
