+++ klugee-main-thing/database/migrations/2021_11_28_051734_create_teach_programs_table.php
@@ -15,7 +15,7 @@ class CreateTeachProgramsTable extends Migration
     {
         Schema::create('teach_programs', function (Blueprint $table) {
             $table->id();
-            $table->integer("id_teacher");
+            $table->bigInteger("id_teacher");
             $table->string("program_name");
             $table->timestamps();
         });
