+++ klugee-main-thing/database/migrations/2021_11_28_054233_create_student_programs_table.php
@@ -15,7 +15,7 @@ class CreateStudentProgramsTable extends Migration
     {
         Schema::create('student_programs', function (Blueprint $table) {
             $table->id();
-            $table->integer("id_student");
+            $table->bigInteger("id_student");
             $table->string("program");
             $table->timestamps();
         });
