+++ klugee-main-thing/database/migrations/2022_01_04_173048_create_teach_methods_table.php
@@ -15,7 +15,7 @@ class CreateTeachMethodsTable extends Migration
     {
         Schema::create('teach_methods', function (Blueprint $table) {
             $table->id();
-            $table->integer("id_teacher");
+            $table->bigInteger("id_teacher");
             $table->string("method");
             $table->timestamps();
         });
